+++ CAR_TRACKING/for_use_GPU.h
@@ -2,17 +2,18 @@
 #define _CUDA_H
 #include <cuda.h>
 #endif
+#include "switch_float.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
     
 struct thread_data {
-    double *A;
-    double *B;
-    double *C;
-    double *F;
-    double *T;
+    FLOAT *A;
+    FLOAT *B;
+    FLOAT *C;
+    FLOAT *F;
+    FLOAT *T;
     int A_dims[3];
     int B_dims[3];
     int C_dims[2];
@@ -39,8 +40,7 @@ extern void clean_cuda(void);
 extern char *conv(unsigned int res);
 
 /* function for GPU execution correspond to fconvsMT */
-//extern double ***fconvsMT_GPU(CUdeviceptr featp2_dev, double **filter,int *sym_info,int start,int end,int **A_SIZE, CUdeviceptr A_SIZE_dev, int **B_SIZE,int **M_size_array, int L_MAX, int interval, int *FSIZE, int padx, int pady, int max_X, int max_Y, int calc_flag);
-extern double ***fconvsMT_GPU(CUdeviceptr featp2_dev, double **filter,int *sym_info,int start,int end,int *A_SIZE, CUdeviceptr A_SIZE_dev, int **B_SIZE,int **M_size_array, int L_MAX, int interval, int *FSIZE, int padx, int pady, int max_X, int max_Y, int calc_flag);
+extern FLOAT ***fconvsMT_GPU(CUdeviceptr featp2_dev, FLOAT **filter,int *sym_info,int start,int end,int *A_SIZE, CUdeviceptr A_SIZE_dev, int **B_SIZE,int **M_size_array, int L_MAX, int interval, int *FSIZE, int padx, int pady, int max_X, int max_Y, int calc_flag);
     
 /* definition of calc_flag */ 
 #define ROOT 0
@@ -51,6 +51,7 @@ extern double ***fconvsMT_GPU(CUdeviceptr featp2_dev, double **filter,int *sym_i
 //#define ORIGINAL
 
 //#define SEPARETE_MEM
+
 
     
 #ifdef __cplusplus
